LayoutTests/imported/w3c/web-platform-tests/XMLHttpRequest/access-control-basic-post-with-non-cors-safelisted-content-type-expected.txt

CONSOLE MESSAGE: line 33: XMLHttpRequest cannot load http://127.0.0.1:8800/XMLHttpRequest/resources/access-control-basic-options-not-supported.py. Preflight response is not successful

PASS Same-origin request with non-safelisted content type succeeds 
PASS CORS request with non-safelisted content type sends preflight and fails 
